Remove multiparty for form control
we had form control and validation using the multiparty module
and did the form parsing on the candidatos.js file.

now the form is controlled through a javascript file on the
candidate creation page called new.js

with the old way of doing it the fields returned values as arrays
and we had to take the first index of each field to validate them.