Add middleware validator for Candidato
we now use a validator middleware to ensure that we are receiving
a valid Candidato. This will be used when updating, editing or
inserting a Candidato, and separates the logic from our routes. If
the Candidato is not valid, the proper error is handled by passing
the error to Next()